+++ bvconversations/src/main/java/com/bazaarvoice/bvandroidsdk/ReivewSummaryResponse.java
@@ -1,0 +1,5 @@
+package com.bazaarvoice.bvandroidsdk;
+
+public class ReivewSummaryResponse extends  IncludesResponse<ReivewSummary, ConversationsIncludeProduct>{
+}
+